Shell Script to Enhance the Calendar to Accept MM, MMM, YYYY
In the script, if the user enters one argument then we will check whether it is greater than 12, if it is greater than 12 then the argument will be considered as a year else argument will be considered as a month. If no argument is given, then “Invalid arguments is displayed.”. If 2 arguments are given then it will denote a month and a year. 

To run the shell script use:

./cal.sh <argument 1> <argument 2>